Implementasi slug

+++ app/Series.php
@@ -4,10 +4,22 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 use Storage;
+use Cviebrock\EloquentSluggable\Sluggable;
 
 class Series extends Model
 {
     protected $table = 'series';
+
+    use Sluggable;
+
+    public function sluggable()
+    {
+        return [
+        'slug' => [
+          'source' => 'title'
+        ]
+      ];
+    }
 
     public function genre()
     {
@@ -28,11 +40,12 @@ class Series extends Model
 
             Storage::delete('public/'.$episode->cover); // hapus cover sendiri
 
-            foreach ($episode->episode as $kampret) {
-                $arr[] = 'public/'.$kampret->cover;
+            if (count($episode->episode)) { // cek jika ada episode nya
+                foreach ($episode->episode as $kampret) {
+                    $arr[] = 'public/'.$kampret->cover;
+                }
+                Storage::delete($arr); // hapus cover yang ada di episode juga
             }
-            Storage::delete($arr); // hapus cover yang ada di episode juga
-
             $episode->episode()->delete();
         });
     }
